Created a separate project of component14 and component15 for JRE <= 1.4 and JRE > 5 respectively under jsf-flex-framework project.
Separated the implementation of the task-runners from jsf-flex-framework core project into runnerImpl project [i.e. Ant's implementation of flexTaskRunner is held within jsf-flex-framework\runnerImpl\flexTaskRunnerImpl\antFlexTaskRunnerImpl]

Created the jsf-flex-build-plugIn to generate the replaceMappingXML for component14 during the build process and deleted all the replaceMappingXML from the component14 project. The XML files will be created by reading the documentation of the Java files. component15 will be read from annotation [have yet to test it]. So component14 and component15 will have an implementation of _AnnotationDocletParser which will be created for jsf-flex-framework core during run time.

All the templates except for added bodyContent [i.e. MXMLApplication and MXMLScript] have been deleted and will depend on the implementation of _FileManipulatorTaskRunner to create these templates [i.e. default is VelocityFileManipulatorTaskRunnerImpl].

Determination of the implementation _CommonTaskRunner, _FileManipulatorTaskRunner, and _FlexTaskRunner will be made during the build process. Implementation of _RunnerFactory will dictate which class to instantiate during run time and one can check out the DefaultRunnerFactoryImpl for specs. For ref, the property file's values under resources are populated during the build process. 

Added the task to create SWC files so to link the system's code dynamically and to lighten the file size of each SWF files.

All the changes were made with intention to separate the implementation from interfaces and to follow the IOC design pattern.

Will create various Wikipages explaining much of this.