Initial refactoring of the code to have the rendering done by the Renderers. So removed component14 and component15
project and created renderKit14 and renderKit15 projects. There exists many tasks to be completed and have been noted with 

TODO + HACK comments. For example, MXMLResponseWriterImpl extends from HtmlResponseWriterImpl for the base implementation, 

while adding additional methods for the JSF Flex project. In the future, one should consider implementing these methods 

specifically for the JSF Flex project. MXMLResponseStateManager extends from HtmlResponseStateManager, once again this 

should be implemented specifically for the JSF Flex project. 

One thing to note is that MXMLRenderKitImplWrapper, which is the renderKit for "MXML_BASIC", allows other RenderKits [i.e. 

"HTML_BASIC"] to be used by storing them within List, but the renderKit factory must be of MXMLRenderKitFactoryImpl.

Much TODO, but hopefully will be able to take care of these matters in the later weeks.